Fixed borkness after .password is used
This is indeed an ugly fix caused by previous bad design.
At the beginning the purpose was to never have socket.db loaded in memory
unless it is really necessary for an ongoing operation. This is important
because otherwise the password, while encrypted, is loaded in memory.

Fixing this will mean having local copies of whatever non-sensitive socket.db
data we might need (at least socket.rank = socket.db.rank).