We're now officially an universe (Ref: #2 )
We're now officially an universe, and the good news is, we are without
depending on nodejs >= 0.12.

The bad news is... there's a lot to do still, and users at this point won't
notice the difference.

https://github.com/marado/TalkerNode/issues/2 has a comprehensive list on what
absolutely needs to be done in order for users to percieve TalkerNode as a
Virtual World.