Fix dependencies versions
I should have read the documentation about versioning before assuming
its syntax.

^version means "compatible with version", but there's no mandatory
versioning scheme, which makes the compatibilty assumptions weak, and
in some cases downright wrong.

In particular:
* dirty is known to have broken compatibility in the past even between
  minor versions, so we can't assume anything in that regard. We know
  it to work in 0.9.9 <= versions <= 1.1.0 (the latest at this
  moment), but maybe the risk of accepting future versions is
  outweighted by the possibility of them not breaking
  retro-compatibility, so we'll risk and go to >= 0.9.9;
* Nodiverse always tries to maintain retro-compatibility, so we just
  want it to be "at least" the version which provides all the
  functionality we're assuming exists (in thise case, >= 0.2.0);
* password-strength has only one release for now, so we shouldn't
  assume any kind of versioning scheme there. However, we're going for
  an optimistic approach by now, as we did with dirty, and assume
  TalkerNode works well with any password-strength version.

More on Semantic Versioning:
While we know that dirty hasn't (until/including the release of 1.0.0)
adhere to Semantic Versioning, it doesn't mean it won't.
For password-strength, we know it didn't adhere to adhere to Semantic
Versioning (by starting with 0.0.1 instead of 0.1.0), but once again,
that doesn't mean it won't.
As for Nodiverse, as the current maintainer of that library, I'm quite
sure it is *not* adhering to Semantic Versioning, and not planning to
do so in the future.
In any case, we could have gone further than the pessimistic approach
and stick with a particular version, at least for the packages we
don't closely control (dirty and password-strength).
I here opted not to do so, maybe an habit flaw (in particular,
following the traditions from package and dependency management on
projects like Debian). I'll be open to accept arguments against this.