Build your own Universe by .dig'ging it!
.dig <direction> <name> will create passage from here towards
<direction> to a new place called <name>.

Note that this dependency on Nodiverse 0.2.1 also means that,
if you have an Universe generated by an older version, you'll
have to ditch it and create a new one (or you can create a new
one, edit the .db file, and copy the new "opposites" object to
the old one).

.dig needs at least one enhancement, not implemented yet: if you have
a place north of here but no passage from here to there, you should
be able to .dig that passage, but that isn't possible yet.

Ref.: #32