removed yarn reference on README
It seems that nowadays the fashionable thing to do is to either support npm or
yarn, but not both. While I'm not one to pay much attention to fashion, having
a pachage-lock.json seems to be enough to make yarn misbehave, even with
yarn import. Let's keep things simple and just support one... and between npm
and yarn, let's stick with the more popular npm.