laying down the groundwork for a pager system
This is a first approach and step towards an implementation of #77 -- a pager
system that allows an user do define they want to recieve their messages a few
lines at a time.

At this moment I am not sure how important is this feature, as TalkerNode does
not send (often) walls of text in one message only, but instead tends to send a
few lines each time, which should solve most of client issues.

On the other hand, if we wish to pursue this, we should probably think about
aggregating everything that is going to be sent to a particular socket (maybe
setting some sort of timeout, and then send it all together). As most of the
messages are origininated in reaction to commands, one other possibility would
be to not have the commands send to a socket directly, but instead to a (number
of) buffer(s) that would be dispatched towards the relevant sockets only when
the command execution ends.