0.6.19: write an auth.log
Each new connection to a TalkerNode server now generates an entry in
the auth.log file, in the following format:

```
2020-04-18T17:08:54.448Z ::ffff:192.168.1.68 connected with port 39408
2020-04-18T17:08:58.243Z ::ffff:127.0.0.1 connected with port 36846
```

As you can see, the timestamps are in ISO format.

The log file is not rotated, and there is no anti-spam mechanism, but
this will be enough to let tools like logrotate and fail2ban to do
their jobs properly.

Closes: #68