0.6.21: deal better with closed connections
When a TCP server is closed, two FINs must occur, and one would expect
that when the server sends a FIN, the client will reply with its own
FIN. However, while that is the commonplace, the connection stays
estabilished, and a client is allowed to send data, and the server
will recieve it, even after a FIN was sent from the server, while the
client doesn't send a FIN of its own.

While we do not need to take particular care about these TCP details,
we should be aware that we can receive data from a (seemingly) closed
connection. And since when we receive data, we often react by sending
data in reply -- and since that there is no point in trying to add
complexity to that -- the best approach to deal with this seems to be
to accept that, sometimes, we will try to send data to sockets we
should know not to. Thus, this patch simply wraps all socket writting
attempts in try-catch clauses, and blindly accepts that they might
fail. Nothing in particular needs to be done in those cases, anyway.

Closes: #95