loadCommands: return the output as a string
Previously, loadCommands did a number of console.log's. Now, it
returns formatted output, so the caller can decide to present or not
that output, in any manner.

Currently, the console keeps going the same logs, when 'rc' is invoked
from it, or at starting time, but this output is now also being sent
through the talker to the invoker of .rc , if it is executed from the
command instead.

Also, if the function fires up an exception, it will only not be
caught at start time, and it will tell, in it, which command failed to
load and caused the exception.