0.6.26: bugfix: validate that the sockets are open
Specially due to the asyncronous nature of what we're doing here,
we should not assume that a socket object is at any point still
open.

For now, we at least validate, in the cycles were we check is a
socket corresponds to a logged in user, that the socket is open.

We also log an error instead of crashing, if we try to write an
exit message to someone and fail to do so.

Closes: #115