0.6.29: stop filtering strings starting with "!$
The "IAC" handling that is coded into TalkerNode is absurdly crude,
and was making a few number of assumptions. Amongst them, it was -
just in case - filtering strings that, after some cleaning, started
with the characters ", $ or !, as they could also correspond to IAC
commands.

This new version simply removes that filtering, allowing for strings
starting with those characters to be said. However, this also means
that we might not be filtering some stuff that we should... This
hopefully doesn't make fixing #34 more urgent, but it is possible.

Closes: #120