depend on strip-ansi < 7
`npm` irks me.

Here's a problem. There's a package that says:

```
  "engines": {
    "node": ">=8.3.0"
  },
  "dependencies": {
    "anotherpackage": ""
  },

```

anotherpackage has two versions. Version 1 says
```
  "engines": {
    "node": ">=8"
  },
```
but version 2 says
```
  "engines": {
    "node": ">=12"
  },
```

**Obviously**, or so I thought, when you npm install the package in question,
and you're running node 10, for instance, it would install the dependencies, ie.
the `anotherpackage`, but, since version 2 depends on a version of node we are
not running with, it and version 1 satisfies the dependency, it should install
version 1. Unfortunately, instead, it issues a warning... but still installs
version 2, and we end up with a broken installation:

```
npm WARN notsup Unsupported engine for anotherpackage@2: wanted: {"node":">=12"} (current: {"node":"10.19.0","npm":"6.14.4"})
npm WARN notsup Not compatible with your version of node/npm: anotherpackage@2
```

While considering this an npm bug (and, to be honest, I'm not sure if a more
recent version of npm doesn't already react better to this scenario), I'm
currently changing TalkerNode's dependency on strip-ansi, forcing to use an old
version, and avoiding users of old node or npm versions to have to suffer due
to this miscalculation of the dependency tree.

Depending on an old depency version is never a good idea, and while this is the
temporary 'fix', a more resilient solution will be to bump up the node engine
minimum version. However, I'm still not ready to take that decision (and drop
support to users of those versions).