Depend on mathjs < 9.4.4
Not unlike what happened in ca64fcd0704f024b5d55290d57682d5888593dd0 ,
we must compensate for the deficiencies of npm.

In this case, mathjs 9.4.4 announces support only for node >= 12.
While it actually works on Node 10[1], we don't know for how long, and
so we're depending specificly on a version older than 9.4.4.

We will need to take this into account when dealing with #126 .

[1] https://github.com/josdejong/mathjs/issues/2258