0.6.37: Deal with user-input color
Now that users can ~BGinput color~RS, we should have a different care
about everything that is user input. For now, this patch makes it so
that:

* descs with color do not ruin .who anymore
* rank names with color do not ruin .who anymore
* texts with color don't bleed out elsewhere

There is probably plenty more where user-input color codes can mess up
with current talker functionality: a thorough investigation of the use
of color with each command should be done.