0.6.38: depend on chalk <= 4
chalk 5 is not retrocompatible (see #137).

For now we are depending on chalk <=4 to fix that, and
Nodiverse >= 0.4.1 (which also depends on chalk <= 4).

In the future we might try to support more recent versions of chalk
too, but that will mean bumping node's dependency to
`"node": "^12.20.0 || ^14.13.1 || >=16.0.0"`, and as we can see in
issue #126 we're still not ready to stop supporting node 10.