0.6.40: commands - fix direct writes to sockets
Several commands* had direct writes to sockets instead of using the
command_access helper that, amongst other things, sanitizes what is
written and takes care of the color codes and relevant escapes.

We should do something to force a stricter access (or restriction)
from commands to the sockets (see #20) but for now we're just fixing
all the commands that where using those writes, and turning them into
sendData calls.

As an extra, some of the commands' output has been 'sanitized' by
adding an ~RS at the end of the message (avoiding color leaks).

This fixes some of the comments in 135#issuecomment-1030891270 .

* demote, emote, go, promote, say, sayto, semote, shout, shoutto, tell